updated processing functions and reorganization libft

+++ libft/ft_strarrdel.c
@@ -6,13 +6,13 @@
 /*   By: svan-der <svan-der@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/15 16:40:10 by svan-der       #+#    #+#                */
-/*   Updated: 2019/05/08 18:41:09 by svan-der      ########   odam.nl         */
+/*   Updated: 2019/10/15 14:04:13 by svan-der      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "includes/libft.h"
 
-void		ft_strarrdel(char ***str, int size)
+void		ft_strarrdel(char ***str, size_t size)
 {
 	char	**ar;
 	size_t	i;
